requirements.txt
pandas~=2.2.0
matplotlib>=3.9.2
python-dateutil~=2.8.2
numpy>=2.1.2